+++ lib/src/theme/themes/default_theme_variant.dart
@@ -229,7 +229,10 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
         color: colorScheme.popover,
         shadows: ShadShadows.md,
       ),
-      anchor: const ShadAnchorAuto(preferBelow: false),
+      anchor: const ShadAnchorAuto(
+        followerAnchor: Alignment.topCenter,
+        targetAnchor: Alignment.topCenter,
+      ),
       duration: Animate.defaultDuration,
       reverseDuration: Duration.zero,
     );
@@ -262,7 +265,7 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
           width: 1,
         ),
       ),
-      anchor: const ShadAnchorAuto(verticalOffset: 24),
+      anchor: const ShadAnchorAuto(),
     );
   }
 
@@ -938,7 +941,11 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
       radius: radius,
       padding: const EdgeInsets.all(4),
       border: ShadBorder.all(color: colorScheme.border, width: 1),
-      anchor: const ShadAnchor(offset: Offset(-8, 8)),
+      anchor: const ShadAnchorAuto(
+        offset: Offset(-8, 8),
+        followerAnchor: Alignment.bottomRight,
+        targetAnchor: Alignment.bottomLeft,
+      ),
       buttonHeight: 32,
       buttonVariant: ShadButtonVariant.ghost,
       buttonSelectedBackgroundColor: colorScheme.accent,