+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - **FIX**: Expose `ShadMouseCursorProvider`.
 - **FIX**: `ShadMenubar` anchor.
+- **FIX**: `ShadBadge` should not enter the gesture arena if the `onPressed` callback is not provided.
+- **FEAT**: Add `cursor` to `ShadBadge`, defaults to `SystemMouseCursors.click` if `onPressed` is provided.
 
 ## 0.22.5